backend/pyproject.toml
[project]
name = "jobuddy-backend"
version = "0.1.0"
description = "Multi-agent AI job application automation system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "crewai>=0.86.0",
    "crewai-tools>=0.14.0",
    "openai>=1.50.0",
    "playwright>=1.49.0",
    "python-docx>=1.1.0",
    "docxtpl>=0.18.0",
    "websockets>=13.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
